Screenshot folder of logs
Because we used UnityCollaborate and PlasticSCM to collaborate and work with each other, most of the work that we did will not be registered on GitHub. This folder demonstrates everyone's contribution to the project through a series of screenshots. We can provide our logs on each of our computers when asked.